Added better error checking for HEAD redirects resolution, alongside a general tidy-up.
